Room demo

+++ src/com/rehab/animation/Drawable.java
@@ -14,7 +14,7 @@ public interface Drawable {
 	
 	void setSprite(Sprite s); 
 	
-	double getZ(); 
+	int getZ();
 	
 	void onClick();
